# radio-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Radiology API",
        "description": "Appointment booking application for a radiology department",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local server"
        },
        {
            "url": "https://radiologie-rdv.com/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/creneaux/{service_id}": {
            "get": {
                "tags": [
                    "Créneaux"
                ],
                "summary": "Lister les créneaux disponibles pour un service donné",
                "description": "Retourne les créneaux disponibles, avec un traitement spécial si le paramètre urgent=1 est présent.",
                "operationId": "bcacd17f052cc3a2a35e7749b806ad10",
                "parameters": [
                    {
                        "name": "service_id",
                        "in": "path",
                        "description": "Identifiant du service",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "urgent",
                        "in": "query",
                        "description": "Si égal à 1, filtre les créneaux pour les 24 prochaines heures",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des créneaux disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "time": {
                                                "type": "string",
                                                "example": "09:00"
                                            },
                                            "end_time": {
                                                "type": "string",
                                                "example": "10:00"
                                            },
                                            "service_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Service non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/creneaux-disponibles/{serviceId}": {
            "get": {
                "tags": [
                    "Créneaux"
                ],
                "summary": "Lister les créneaux disponibles pour un service",
                "description": "Retourne les créneaux disponibles filtrés selon l'urgence (paramètre urgent=1).",
                "operationId": "73384126fa31411e6f915cba166a61ab",
                "parameters": [
                    {
                        "name": "serviceId",
                        "in": "path",
                        "description": "ID du service médical",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "urgent",
                        "in": "query",
                        "description": "Si 1, affiche uniquement les créneaux des prochaines 24h",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des créneaux disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "time": {
                                                "type": "string",
                                                "example": "09:00"
                                            },
                                            "service_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/medecin/dashboard": {
            "get": {
                "tags": [
                    "Médecin"
                ],
                "summary": "Afficher le tableau de bord du médecin",
                "description": "Retourne la vue du tableau de bord pour un médecin authentifié.",
                "operationId": "b0337369393f6d0122e845da365ed7b2",
                "responses": {
                    "200": {
                        "description": "Vue du tableau de bord affichée avec succès."
                    },
                    "401": {
                        "description": "Utilisateur non authentifié."
                    },
                    "403": {
                        "description": "Accès refusé. L'utilisateur n'est pas médecin."
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "Profil"
                ],
                "summary": "Afficher les informations du profil de l'utilisateur",
                "description": "Retourne les données du profil de l'utilisateur authentifié.",
                "operationId": "8cfcc99d8d1aa366314cd8abf9cd9c9a",
                "responses": {
                    "200": {
                        "description": "Formulaire du profil affiché avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profil"
                ],
                "summary": "Supprimer le compte utilisateur",
                "description": "Supprime définitivement le compte de l'utilisateur après vérification du mot de passe.",
                "operationId": "1d6a5c6489e35a763aa97ebef12863ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "motdepasse123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Utilisateur déconnecté et redirigé vers la page d'accueil"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Mot de passe incorrect ou validation échouée"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Profil"
                ],
                "summary": "Mettre à jour les informations du profil",
                "description": "Met à jour les informations du profil de l'utilisateur authentifié. Si l'email change, la vérification est réinitialisée.",
                "operationId": "354f5b23296bed4036564563e7423f43",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ayoub Berhili"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ayoub@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Profil mis à jour et redirigé vers la page de profil"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sessionAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token for authentication",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    },
...
}